requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.21
albumentations==1.4.23
altgraph==0.17.4
annotated-types==0.7.0
astunparse==1.6.3
basicsr==1.4.2
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
coloredlogs==15.0.1
comtypes==1.4.8
contourpy==1.3.1
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.15
cycler==0.12.1
Cython==3.0.11
darkdetect==0.8.0
easydict==1.13
eval_type_backport==0.2.2
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.55.3
future==1.0.0
gast==0.6.0
gdown==5.2.0
gfpgan==1.3.8
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_resources==6.4.5
insightface==0.7.3
Jinja2==3.1.5
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
macholib==1.16.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
modulegraph==0.19.6
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
onnx==1.16.0
onnxruntime-silicon==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.10.0.84
opennsfw2==0.10.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
Pillow==9.5.0
platformdirs==4.3.6
prettytable==3.12.0
protobuf==4.23.2
psutil==5.9.8
py2app==0.28.8
pydantic==2.10.4
pydantic_core==2.27.2
pydeps==2.0.1
Pygments==2.18.0
pygrabber==0.2
pyparsing==3.2.0
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
simsimd==6.2.1
six==1.17.0
soupsieve==2.6
stdlib-list==0.11.0
stringzilla==3.11.2
sympy==1.13.3
tb-nightly==2.19.0a20241222
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tk==0.1.0
tkinterdnd2==0.4.2
tomli==2.2.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
yapf==0.43.0
